17173 mdb: check sign expansion Reviewed by: Gordon Ross <gordon.w.ross@gmail.com> Reviewed by: Andy Fiddaman <illumos@fiddaman.net> Approved by: Dan McDonald <danmcd@edgecast.io>

+++ usr/src/cmd/mdb/common/mdb/mdb_print.c
@@ -2727,13 +2727,13 @@ printf_signed(mdb_ctf_id_t id, uintptr_t addr, ulong_t off, char *fmt,
 
 	switch (size) {
 	case sizeof (uint8_t):
-		mdb_printf(fmt, (uint64_t)(sign ? u.i1 : u.ui1));
+		mdb_printf(fmt, (uint64_t)(sign ? (uint64_t)u.i1 : u.ui1));
 		break;
 	case sizeof (uint16_t):
-		mdb_printf(fmt, (uint64_t)(sign ? u.i2 : u.ui2));
+		mdb_printf(fmt, (uint64_t)(sign ? (uint64_t)u.i2 : u.ui2));
 		break;
 	case sizeof (uint32_t):
-		mdb_printf(fmt, (uint64_t)(sign ? u.i4 : u.ui4));
+		mdb_printf(fmt, (uint64_t)(sign ? (uint64_t)u.i4 : u.ui4));
 		break;
 	case sizeof (uint64_t):
 		mdb_printf(fmt, (uint64_t)(sign ? u.i8 : u.ui8));
